feat: add web worker for markdown 2 kanban
It's important for us to update the UI on every keystroke
for the UI to feel fluid.

It takes 50ms for about 100 tasks being parsed,
which is too long and is very noticeable when typing.

Instead, we can use this web worker to offload the AST parsing to a
separate thread that doesn't block the main UI rendering.